requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioresponses==0.7.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-lru==2.0.4
attrs==24.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.6
flake8==7.1.1
frozenlist==1.5.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pycodestyle==2.12.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pyflakes==3.2.0
pytest==8.3.4
pytest-asyncio==0.25.1
python-dotenv==1.0.1
redis==5.2.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
